Moved close() method to SubscriberFactory to handle life cycle event
